src/requirements/base.txt
certifi==2021.5.30
celery==5.1.2
cryptography==3.3.2
Django==3.2.7
django-anymail==8.4
django-countries==7.2.1
django-ckeditor==6.1.0
django-colorful==1.3
django-cors-headers==3.9.0
django-filter==2.4.0
django-modeltranslation==0.17.3
django-money==2.1
django-mptt==0.13.4
django-mptt-admin==2.1.0
django-nested-admin==3.3.3
django-optimized-image==0.3.0
django-split-settings==1.1.0
djangorestframework-simplejwt==4.8.0
djangorestframework==3.12.4
graphene-django==2.15.0
idna==3.2
Markdown==3.3.4
oauthlib==3.1.1
phonenumbers==8.12.33
Pillow==8.3.2
psycopg2==2.9.1
python-decouple==3.4
python3-openid==3.2.0
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.0
sentry-sdk==1.4.3
shortuuid==1.0.1
social-auth-app-django==5.0.0
social-auth-core==4.1.0
six==1.16.0
social-auth-app-django==5.0.0
sqlparse==0.4.2
urllib3==1.26.7
django-registration==3.2
django-redis==5.0.0
drf-yasg2==1.19.4
django-robots==4.0
notifiers==1.2.1